pyproject.toml
[project]
name = "sg-ai-ctf-25"
version = "0.1.0"
description = "All challenges for SG-AI CTF 2025"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-image>=0.25.2",
    "selenium>=4.36.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
